Background color not 100% of document height
Since introducing embedded sections the background no -longer stretches
to 100% of the page height.  This has been remedied by introducing a
drop shadow on the `.waltz-body` element.

A neater solution may have been to use
```
height: fill-available  // with -moz and -webpack variants
```

However this is not supported under IE.

Bonus:
- tightened up parameter declarations in routes

#CTCTOWALTZ-1443
#4319