Attestations: closing instances fails for logical and physical flows
Root cause: assumption that the run has an attested_entity_id - this is only for measurable attestations (i.e the category id).
This in turn caused the sql to not match on the run and therefore it creates a new run/instance on the fly rather than closing the existing one.

#5661