Assessment support for measurables
Added qualifier support to table.
Added findByEntityReference which does a lookup if passed a measurable in order to fill in the qualifier.
Resorted the list of entity kinds supported by assessments.
Added assessment section to measurable page sections.
Added assessment sub-section to measurable header.
Svelte Store will remove the qualifier if the def is not a measurable

#6125