Better back btn support
We seem to have a problem with window.history.replaceState, it writes entries in the history which seem to block the back button from working correctly.

The code is in section-store and is only really needed to update the query param with a list of open sections. In turn this is only needed so the url can be sent to others and the appropriate sections will open.

After some experimentation it seems as if we can remove that feature and instead us a more straightforward sharing mechanism, where a shareable url is copied to the clip board.

#6657